fix: move json back to the project root
